+++ CHANGELOG.md
@@ -5,6 +5,8 @@ the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## Unreleased
 
+- Fixed: Context owning model now always has at least one context. This fixes the issue where model becomes context
+  owner after creation, no context was available so no fragments could be added.
 - Added: Show a new tree breadcrumbs column when in Table filtering/sorting mode. This can be (de)selected as column
 
 ## [0.10.20] - 2026-03-17